refactor: move Constants to resources folder
